Updated to Unicode Ver6.3. Added uniNamesList_name(), uniNamesList_annot()
Update 6.3 does not affect Nameslist.txt, which remains at version 6.2.
The French NamesList does not appear to have any further updates after
version 5.0, so we really only have an English NamesList.txt to work with.

It is easier to find this library using AC_CHECK_LIB or AC_SEARCH_LIB if
there is an executable function inside the library itself to check for.

These two functions return a pointer to const char string data if a name
or annotation exists, otherwise return NULL if there isn't one to return.