Explain 3 constants in more detail for developers that use libuninameslist
Libuninameslist is based on the data found in the Nameslist.txt file from
www.unicode.org which can/will change with newer/later updates. To allow
a program to be prepared for future updates of libuninameslist without the
need to modify your program, you need to see UnicodeBlock[] and not see it
as UnicodeBlock[some_constant]. For example Nameslist.txt version 5.0 had
168 blocks, while Nameslist.txt 6.2 has 234, meanwhile the preliminary 6.3
which is not official yet, has more blocks. Example code can be found in
FontForge "unicode_block_check(int block_i)" to find the (last) Nth block.

This may help resolve Debian Bug #688387 "Missing entries in UnicodeBlock"