ldconfig picked-up in debian lintian therefore might be a problem
Reading various buglists and issues seems some distros may need you to
run ldconfig if you plan to install another program immediately after
building LibUninamesList and doing a make install (without rebooting).
There are some suggestions to use install-exec-hook in Makefile.am,
but this causes problems with some OSes, or if you install as non-'su'
so the easiest solution is to run ldconfig manually for installers
that need this, and skip it, for those that don't need it.