Build failure with clang - bugfix
This fixes a bug with clang running on MSYS2. Windows expects to have
import and export declarations to reduce the size of the lookup tables.
gcc >4 introduced visibility, but a lot of code is not using this often
and it becomes a problem for code in win32/win64 due to lack of this.
CLANG for MSYS2 quietly injects the import/export declarations itself,
so it can make use of gcc code so we need to turn-off this feature if
compiling with CLANG on MSYS2 to avoid re-inserting this information 2x