Unicode 15.0 includes a number of changes and 4,489 new characters, including another major extension of CJK unified ideographs. A number of the Unicode Standard Annexes have significant modifications for Unicode 15.0. Two new scripts have been added, and there are also 20 additional emoji characters in Unicode 15.0.
A. Summary

Unicode 15.0 adds 4489 characters, for a total of 149,186 characters. These additions include 2 new scripts, for a total of 161 scripts, as well as 20 new emoji characters.

unicode-15-0-beta-review/

NamesList.txt 15.0_Beta for anyone interested in testing the next version.

If you have any Unicode bug reports or improvements, please report them at
https://home.unicode.org/unicode-15-0-beta-review/
The review period for 15.0 Beta is before 2022jul12.

This is marked as 14.9 for testing purposes.